Refactor WhatsApp routes into dedicated router
